requirement.txt
absl-py==2.3.1
altair==5.5.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.4.0
Automat==25.4.16
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comm==0.2.3
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.3
cssselect==1.3.0
cycler==0.12.1
DateTime==5.5
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.8.0
dotenv==0.9.9
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
fsspec==2025.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itemadapter==0.12.2
itemloaders==1.3.2
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
keras==3.12.0
kiwisolver==1.4.9
lark==1.3.1
libclang==18.1.1
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.3
mpmath==1.3.0
mysql-connector-python==9.5.0
namex==0.1.0
narwhals==2.10.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.3.4
opt_einsum==3.4.0
optree==0.17.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.0
plotly==6.3.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
Protego==0.5.0
protobuf==6.33.0
psutil==7.1.2
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydeck==0.9.1
PyDispatcher==2.0.7
Pygments==2.19.2
pymongo==4.15.3
pyOpenSSL==25.3.0
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
queuelib==1.8.0
referencing==0.37.0
requests==2.32.5
requests-file==3.0.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.28.0
scikit-learn==1.7.2
scipy==1.16.3
Scrapy==2.13.3
selenium==4.38.0
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
stack-data==0.6.3
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tldextract==5.3.0
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.31.0
trio-websocket==0.12.2
Twisted==25.5.0
typing_extensions==4.15.0
tzdata==2025.2
unicode==2.9
Unidecode==1.4.0
uri-template==1.3.0
urllib3==2.5.0
w3lib==2.3.1
watchdog==6.0.0
wcwidth==0.2.14
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
wsproto==1.2.0
zope.interface==8.0.1
